+++ NEWS.rst
@@ -7,6 +7,8 @@ of every change, see the Git log.
 Latest (head of the master branch)
 ----------------------------------
 * Fixed Python3.x support, broken include statement.
+* Added wurf_waf_unit_test tool
+* Added wurf_protoc & wurf_proto_cxx tools for protobuf support
 
 1.0.0
 -----